+++ accessories/netgear/aiowndr.py
@@ -16,7 +16,7 @@ rxdct = dict(
 	rxYestday=re.compile(r'yesterday_down='+RENUM)
 )
 
-async def traffic(session: aiohttp.ClientSession, user = 'admin',pwd=None, host = '192.168.1.1', semaphore=None):
+async def traffic(session: aiohttp.ClientSession, user = 'admin',pwd=None, host='192.168.1.1', semaphore=None):
 	''' gets traffic.htm page from host, parses page looking for items in rxdct '''
 	#url = 'http://' + host +'/traffic_meter_2nd.htm' 
 	#url = 'http://' + host +'/RST_statistic.htm'
@@ -33,7 +33,7 @@ async def traffic(session: aiohttp.ClientSession, user = 'admin',pwd=None, host 
 		
 	for i in range(2):  # try it twice
 		try:
-			async with semaphore, session.get(url=url, auth=auth, timeout=2.0) as response:
+			async with semaphore, session.get(url=url, auth=auth, timeout=4) as response:
 				stuff = await response.text()
 				if len(stuff)>1 and response.status<400:
 					break  # success
@@ -49,16 +49,16 @@ async def traffic(session: aiohttp.ClientSession, user = 'admin',pwd=None, host 
 				resp[itm] = float(grp.replace(',',''))
 				logger.debug('%s mtch.grp=%s=%f spn=%s' % (itm,grp,resp[itm], mtch.span()))
 			else:
-				logger.warning('no match with %s in %s for %s' % (rx, len(stuff), itm))
+				logger.debug('no match with %s in %s for %s' % (rx, len(stuff), itm))
 		if len(resp)==0:
-			logger.warning('wndr nothing found in stuff:%s:=>%s' % (stuff,rxdct))
+			logger.debug('wndr nothing found in stuff:%s:=>%s' % (stuff,rxdct))
 	except Exception as er:
 		logger.error('%d http parse:%s' % (i,er))
 	return resp
 
-async def get_traffic(interval=15, host='192.168.1.1', pwd=None, semaphore=None):
+async def get_traffic(host='192.168.1.1', pwd=None, semaphore=None):
 	""" get traffic from wndr router in Mbytes """
-	await asyncio.sleep(interval)
+	await asyncio.sleep(0.04)
 	#if semaphore is None:
 	#	semaphore = HueBaseDev.Semaphore
 	async with aiohttp.ClientSession() as session: